+++ admin/src/components/BlocksInput/AdvancedSettingsIcons.tsx
@@ -1,0 +1,115 @@
+export const ViewportIcon = () => {
+  return (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="3" y="3" width="5" height="18" rx="1.5" fill="currentColor" />
+      <rect x="10" y="3" width="11" height="18" rx="1.5" fill="currentColor" />
+    </svg>
+  );
+};
+
+export const FontSizeIcon = () => {
+  return (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <text x="2" y="18" fontSize="18" fontFamily="Arial" fill="currentColor">
+        A
+      </text>
+      <text x="14" y="14" fontSize="10" fontFamily="Arial" fill="currentColor">
+        A
+      </text>
+    </svg>
+  );
+};
+
+export const FontLeadingIcon = () => {
+  return (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <text x="4" y="16" fontSize="14" fontFamily="Arial" fill="currentColor">
+        A
+      </text>
+      <path d="M16 6 L16 18" stroke="currentColor" strokeWidth="1.5" />
+      <path d="M14 8 L16 6 L18 8" fill="none" stroke="currentColor" strokeWidth="1.5" />
+      <path d="M14 16 L16 18 L18 16" fill="none" stroke="currentColor" strokeWidth="1.5" />
+    </svg>
+  );
+};
+
+export const FontTrackingIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 20 24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      transform="rotate(90)"
+    >
+      <text
+        x="5.5"
+        y="27"
+        fontSize="14"
+        fontFamily="Arial"
+        fill="currentColor"
+        transform="rotate(-90 4 18)"
+      >
+        A
+      </text>
+      <path d="M16 6 L16 18" stroke="currentColor" strokeWidth="1.5" />
+      <path d="M14 8 L16 6 L18 8" fill="none" stroke="currentColor" strokeWidth="1.5" />
+      <path d="M14 16 L16 18 L18 16" fill="none" stroke="currentColor" strokeWidth="1.5" />
+    </svg>
+  );
+};
+
+export const FontAlignmentIcon = () => {
+  return (
+    <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <rect x="4" y="5" width="16" height="1.5" fill="currentColor" />
+      <rect x="4" y="9" width="12" height="1.5" fill="currentColor" />
+      <rect x="4" y="13" width="16" height="1.5" fill="currentColor" />
+      <rect x="4" y="17" width="10" height="1.5" fill="currentColor" />
+    </svg>
+  );
+};
+
+export const SeparatorSizeIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path d="M 2,4.5 22,4.5 22,8 2,8 2,4.5 Z m 0,5 20,0 0,3 -20,0 0,-3 z m 0,4.5 20,0 0,2.25 -20,0 0,-2.25 z m 0,3.75 20,0 0,1.5 -20,0 0,-1.5 z" />
+    </svg>
+  );
+};
+
+export const SeparatorLengthIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path d="M24 12l-4.5-4.5v3h-15v-3L0 12l4.5 4.5v-3h15v3z" />
+    </svg>
+  );
+};
+
+
+export const SeparatorOrientationIcon = () => {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path d="M11,3H5A2.002,2.002,0,0,0,3,5V19a2.002,2.002,0,0,0,2,2h6a2.002,2.002,0,0,0,2-2V5A2.002,2.002,0,0,0,11,3ZM5,19V5h6l.00146,14Zm14-8H15a1,1,0,0,0,0,2h4v6H15a1,1,0,0,0,0,2h4a2.002,2.002,0,0,0,2-2V13A2.002,2.002,0,0,0,19,11ZM15,7a2.002,2.002,0,0,1,2,2,1,1,0,0,0,2,0,4.00458,4.00458,0,0,0-4-4,1,1,0,0,0,0,2Z" />
+    </svg>
+  );
+};