+++ admin/src/components/BlocksInput/FontModifiersIcons.tsx
@@ -19,8 +19,8 @@ export const Superscript = ({ fill }: FontModifierIconProps) => {
 
   return (
     <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <text x="5" y="17" font-family="Arial" font-size="14" fill={colors[fill as keyof typeof colors]}>X</text>
-      <text x="15" y="10" font-family="Arial" font-size="8" fill={colors[fill as keyof typeof colors]}>1</text>
+      <text x="5" y="17" fontFamily="Arial" fontSize="14" fill={colors[fill as keyof typeof colors]}>X</text>
+      <text x="15" y="10" fontFamily="Arial" fontSize="8" fill={colors[fill as keyof typeof colors]}>1</text>
     </svg>  
   )
 }
@@ -30,8 +30,8 @@ export const Subsscript = ({ fill }: FontModifierIconProps) => {
 
   return (
     <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <text x="5" y="17" font-family="Arial" font-size="14" fill={colors[fill as keyof typeof colors]}>X</text>
-  <text x="15" y="20" font-family="Arial" font-size="8" fill={colors[fill as keyof typeof colors]}>1</text>
+  <text x="5" y="17" fontFamily="Arial" fontSize="14" fill={colors[fill as keyof typeof colors]}>X</text>
+  <text x="15" y="20" fontFamily="Arial" fontSize="8" fill={colors[fill as keyof typeof colors]}>1</text>
 </svg>
 
   )